% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{animate}
\alias{animate}
\title{Animate an S3 object of class IFS}
\usage{
animate(IFS, glebokosc, lkolorow, lfunkcji)
}
\arguments{
\item{IFS, }{object of class IFS}

\item{glebokosc}{depth of fractal}

\item{lkolorow}{number of colors}

\item{lfunkcji}{number of functions}
}
\value{
animated plot
}
\description{
Animate an S3 object of class IFS
}
\examples{

t1<-function(vec){
x<-vec[1]
y<-vec[2]
x<-0.01*x
y<-0.45*y
return(c(x,y))
}

t2<-function(vec){
  x<-vec[1]
  y<-vec[2]
  x<-(-0.01)*x
  y<-(-0.45)*y+0.4
  return(c(x,y))
}

t3<-function(vec){
  x<-vec[1]
  y<-vec[2]
  x<-0.42*x-0.42*y
  y<-0.42*x+0.42*y+0.4
  return(c(x,y))
}

t4<-function(vec){
  x<-vec[1]
  y<-vec[2]
  x<-0.42*x+0.42*y
  y<-(-0.42)*x+0.42*y+0.4
  return(c(x,y))
}

s1<-function(vec){
  x<-vec[1]
  y<-vec[2]
  x<-1/2*x
  y<-1/2*y
  return(c(x,y))
}

s2<-function(vec){
  x<-vec[1]
  y<-vec[2]
  x<-1/2*x+1/2
  y<-1/2*y
  return(c(x,y))
}
s3<-function(vec){
  x<-vec[1]
  y<-vec[2]
  x<-1/2*x+1/4
  y<-1/2*y+3**(1/2)/4
  return(c(x,y))
}

b1<-function(vec){
  x<-vec[1]
  y<-vec[2]
  x<-0.85*x+0.04*y
  y<-(-0.04)*x+0.85*y+1.6
  return(c(x,y))
}

b2<-function(vec){
  x<-vec[1]
  y<-vec[2]
  x<-(-0.15*x)+0.28*y
  y<-0.26*x+0.24*y+0.44
  return(c(x,y))
}

b3<-function(vec){
  x<-vec[1]
  y<-vec[2]
  x<-0.2*x-0.26*y
  y<-0.23*x+0.22*y+1.6
  return(c(x,y))
}

b4<-function(vec){
  x<-vec[1]
  y<-vec[2]
  x<-0
  y<-0.16*y
  return(c(x,y))
}

ss1<-function(vec){
  x<-vec[1]
  y<-vec[2]
  x<-x/3-1/3
  y<-y/3+1/3
  return(c(x,y))
}

ss2<-function(vec){
  x<-vec[1]
  y<-vec[2]
  x<-x/3
  y<-y/3+1/3
  return(c(x,y))
}

ss3<-function(vec){
  x<-vec[1]
  y<-vec[2]
  x<-x/3+1/3
  y<-y/3+1/3
  return(c(x,y))
}

ss4<-function(vec){
  x<-vec[1]
  y<-vec[2]
  x<-x/3+1/3
  y<-y/3
  return(c(x,y))
}

ss5<-function(vec){
  x<-vec[1]
  y<-vec[2]
  x<-x/3+1/3
  y<-y/3-1/3
  return(c(x,y))
}

ss6<-function(vec){
  x<-vec[1]
  y<-vec[2]
  x<-x/3
  y<-y/3-1/3
  return(c(x,y))
}

ss7<-function(vec){
  x<-vec[1]
  y<-vec[2]
  x<-x/3-1/3
  y<-y/3-1/3
  return(c(x,y))
}

ss8<-function(vec){
  x<-vec[1]
  y<-vec[2]
  x<-x/3-1/3
  y<-y/3
  return(c(x,y))
}

triangle<-createIFS(s1,s2,s3)
tree<-createIFS(t1,t2,t3,t4)
barnsley<-createIFS(b1,b2,b3,b4)
square<-createIFS(ss1,ss2,ss3,ss4,ss5,ss6,ss7,ss8)
\dontrun{
animate(triangle,9,3,3)
animate(triangle,9,9,3)
animate(triangle,9,6,3)
animate(tree,8,4,4)
animate(barnsley,20,8,4)
animate(square,5,50,8)
animate(square,5,16,8)
}
}
